kubernetesProtection/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
crowdstrike-falconpy==1.4.0
idna==3.6
requests==2.31.0
#urllib3==2.1.0
